move more functionality to shared dir, update paths
