+++ shared/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "baseUrl": "./",
+        "strict": true,
+        "forceConsistentCasingInFileNames": true,
+        "esModuleInterop": true,
+        "resolveJsonModule": true,
+        "incremental": true,
+        "typeRoots": ["./node_modules/@types/", "../node_modules/@types/", "./types/"]
+    },
+    "include": ["**/*.ts"],
+    "exclude": ["node_modules"]
+}
