+++ shared/types/drop.ts
@@ -2,8 +2,7 @@ import type { BaseContext } from './common';
 import type { EventObject } from 'xstate/lib/types';
 import type Peer from 'peerjs';
 import type { DataConnection } from 'peerjs';
-import { DropEventType } from '@shared/lib/constants';
-import { useDrop } from 'hooks/use-drop';
+import { DropEventType } from 'lib/constants';
 
 export type DropOptions = {
     decryptedAccess?: 'copy' | 'view' | 'both';
@@ -58,5 +57,3 @@ export interface HandshakeCompleteEvent extends DropEvent {
 export interface CompleteEvent extends DropEvent {
     type: DropEventType.Confirm;
 }
-
-export type DropContextValues = ReturnType<typeof useDrop>;
