+++ shared/types/global.d.ts
@@ -2,7 +2,3 @@
 interface Crypto {
     webkitSubtle?: SubtleCrypto;
 }
-
-interface Window {
-    GA_ANALYTICS: boolean;
-}
