+++ web/contexts/DropContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext } from 'react';
 import { useDrop } from 'hooks/use-drop';
-import { DropContextValues } from 'types/drop';
+import { DropContextValues } from 'types/contexts';
 
 const DropContext = createContext<DropContextValues>({} as any);
