+++ web/types/contexts.ts
@@ -1,0 +1,6 @@
+import { useDrop } from 'hooks/use-drop';
+import { useGrab } from 'hooks/use-grab';
+
+export type DropContextValues = ReturnType<typeof useDrop>;
+
+export type GrabContextValues = ReturnType<typeof useGrab>;
